Change subId to int from long
Conflicts:
	src/java/com/android/internal/telephony/SubscriptionController.java

+++ src/java/android/provider/Telephony.java
@@ -362,7 +362,7 @@ public final class Telephony {
          * @return the URI for the new message
          * @hide
          */
-        public static Uri addMessageToUri(long subId, ContentResolver resolver,
+        public static Uri addMessageToUri(int subId, ContentResolver resolver,
                 Uri uri, String address, String body, String subject,
                 Long date, boolean read, boolean deliveryReport) {
             return addMessageToUri(subId, resolver, uri, address, body, subject,
@@ -408,7 +408,7 @@ public final class Telephony {
          * @return the URI for the new message
          * @hide
          */
-        public static Uri addMessageToUri(long subId, ContentResolver resolver,
+        public static Uri addMessageToUri(int subId, ContentResolver resolver,
                 Uri uri, String address, String body, String subject,
                 Long date, boolean read, boolean deliveryReport, long threadId) {
             return addMessageToUri(subId, resolver, uri, address, body, subject,
@@ -571,7 +571,7 @@ public final class Telephony {
              * @return the URI for the new message
              * @hide
              */
-            public static Uri addMessage(long subId, ContentResolver resolver,
+            public static Uri addMessage(int subId, ContentResolver resolver,
                     String address, String body, String subject, Long date, boolean read) {
                 return addMessageToUri(subId, resolver, CONTENT_URI, address, body,
                         subject, date, read, false);
@@ -629,7 +629,7 @@ public final class Telephony {
              * @return the URI for the new message
              * @hide
              */
-            public static Uri addMessage(long subId, ContentResolver resolver,
+            public static Uri addMessage(int subId, ContentResolver resolver,
                     String address, String body, String subject, Long date) {
                 return addMessageToUri(subId, resolver, CONTENT_URI, address, body,
                         subject, date, true, false);
@@ -674,7 +674,7 @@ public final class Telephony {
              * @return the URI for the new message
              * @hide
              */
-            public static Uri addMessage(long subId, ContentResolver resolver,
+            public static Uri addMessage(int subId, ContentResolver resolver,
                     String address, String body, String subject, Long date) {
                 return addMessageToUri(subId, resolver, CONTENT_URI, address, body,
                         subject, date, true, false);
@@ -741,7 +741,7 @@ public final class Telephony {
              * @return the URI for the new message
              * @hide
              */
-            public static Uri addMessage(long subId, ContentResolver resolver,
+            public static Uri addMessage(int subId, ContentResolver resolver,
                     String address, String body, String subject, Long date,
                     boolean deliveryReport, long threadId) {
                 return addMessageToUri(subId, resolver, CONTENT_URI, address, body,
@@ -1189,7 +1189,7 @@ public final class Telephony {
             public static SmsMessage[] getMessagesFromIntent(Intent intent) {
                 Object[] messages = (Object[]) intent.getSerializableExtra("pdus");
                 String format = intent.getStringExtra("format");
-                long subId = intent.getLongExtra(PhoneConstants.SUBSCRIPTION_KEY,
+                int subId = intent.getIntExtra(PhoneConstants.SUBSCRIPTION_KEY,
                         SubscriptionManager.getDefaultSmsSubId());
 
                 Rlog.v(TAG, " getMessagesFromIntent sub_id : " + subId);
