+++ src/java/android/telephony/SmsMessage.java
@@ -104,13 +104,13 @@ public class SmsMessage {
      *
      * @hide
      */
-    private long mSubId = 0;
+    private int mSubId = 0;
 
     /** set Subscription information
      *
      * @hide
      */
-    public void setSubId(long subId) {
+    public void setSubId(int subId) {
         mSubId = subId;
     }
 
@@ -118,7 +118,7 @@ public class SmsMessage {
      *
      * @hide
      */
-    public long getSubId() {
+    public int getSubId() {
         return mSubId;
     }
